Refactor app.ts
- Move the route handlers out of app.ts
- wrap the setup code in a function and move out of app.ts
- update setup code to seed in memory files metadata store
